backend/requirements.txt
dotenv
langchain
langchain-core
langchain-groq
ipykernel
fastapi
pydantic
typing
uvicorn
motor
python-jose
passlib[bcrypt]
httpx